+++ graphics/gui/gui.cpp
@@ -18,8 +18,6 @@ void Gui::update(float delta, Config &c) {
     imgui_io.DeltaTime = delta;
     
     Vec2 mouse_pos = Vec2(Input::mouse().x, Input::mouse().y);
-    if (c.use_shaders)
-        mouse_pos -= c.window_padding;
     
     imgui_io.MousePos = ImVec2(mouse_pos.x, mouse_pos.y);
     imgui_io.MouseDown[0] = Input::down(SDL_BUTTON_LEFT);