+++ graphics/r_camera.cpp
@@ -5,3 +5,15 @@
 #include "r_camera.h"
 
 using namespace Verse;
+
+namespace {
+    Vec2 *camera;
+}
+
+void Graphics::Camera::bind(Vec2 *c) {
+    camera = c;
+}
+
+void Graphics::Camera::move(int &x, int &y) {
+    
+}